+++ 3rdparty/sokol/sokol_gfx.h
@@ -571,7 +571,7 @@ enum {
     SG_MAX_COLOR_ATTACHMENTS = 4,
     SG_MAX_SHADERSTAGE_BUFFERS = 8,
     SG_MAX_SHADERSTAGE_IMAGES = 12,
-    SG_MAX_SHADERSTAGE_IMAGES_CS = 8,
+    SG_MAX_SHADERSTAGE_UAVS = 8,
     SG_MAX_SHADERSTAGE_UBS = 4,
     SG_MAX_UB_MEMBERS = 16,
     SG_MAX_VERTEX_ATTRIBUTES = 16,      /* NOTE: actual max vertex attrs can be less on GLES2, see sg_limits! */
@@ -736,6 +736,7 @@ typedef struct sg_pixelformat_info {
     bool blend;         /* alpha-blending is supported */
     bool msaa;          /* pixel format can be used as MSAA render target */
     bool depth;         /* pixel format is a depth format */
+    bool uav;         /* pixel format can be used as image store (compute-shaders) */
 } sg_pixelformat_info;
 
 /*
@@ -845,6 +846,7 @@ typedef enum sg_buffer_type {
     _SG_BUFFERTYPE_DEFAULT,         /* value 0 reserved for default-init */
     SG_BUFFERTYPE_VERTEXBUFFER,
     SG_BUFFERTYPE_INDEXBUFFER,
+    SG_BUFFERTYPE_RAW,
     _SG_BUFFERTYPE_NUM,
     _SG_BUFFERTYPE_FORCE_U32 = 0x7FFFFFFF
 } sg_buffer_type;
@@ -1079,6 +1081,10 @@ typedef enum sg_uniform_type {
     SG_UNIFORMTYPE_FLOAT3,
     SG_UNIFORMTYPE_FLOAT4,
     SG_UNIFORMTYPE_MAT4,
+    SG_UNIFORMTYPE_INT,
+    SG_UNIFORMTYPE_INT2,
+    SG_UNIFORMTYPE_INT3,
+    SG_UNIFORMTYPE_INT4,
     _SG_UNIFORMTYPE_NUM,
     _SG_UNIFORMTYPE_FORCE_U32 = 0x7FFFFFFF
 } sg_uniform_type;
@@ -1346,6 +1352,8 @@ typedef struct sg_pass_action {
     - 0..1 index buffer offsets
     - 0..N vertex shader stage images
     - 0..N fragment shader stage images
+    - 0..N compute shader stage images
+    - 0..N compute shader storage buffers
 
     The max number of vertex buffer and shader stage images
     are defined by the SG_MAX_SHADERSTAGE_BUFFERS and
@@ -1361,8 +1369,13 @@ typedef struct sg_bindings {
     sg_buffer index_buffer;
     int index_buffer_offset;
     sg_image vs_images[SG_MAX_SHADERSTAGE_IMAGES];
+    sg_buffer vs_buffers[SG_MAX_SHADERSTAGE_BUFFERS];
     sg_image fs_images[SG_MAX_SHADERSTAGE_IMAGES];
+    sg_buffer fs_buffers[SG_MAX_SHADERSTAGE_BUFFERS];
     sg_image cs_images[SG_MAX_SHADERSTAGE_IMAGES];
+    sg_buffer cs_buffers[SG_MAX_SHADERSTAGE_BUFFERS];
+    sg_image cs_image_uavs[SG_MAX_SHADERSTAGE_UAVS];
+    sg_buffer cs_buffer_uavs[SG_MAX_SHADERSTAGE_UAVS];
     uint32_t _end_canary;
 } sg_bindings;
 
@@ -1374,11 +1387,12 @@ typedef struct sg_bindings {
 
     The default configuration is:
 
-    .size:      0       (this *must* be set to a valid size in bytes)
-    .type:      SG_BUFFERTYPE_VERTEXBUFFER
-    .usage:     SG_USAGE_IMMUTABLE
-    .content    0
-    .label      0       (optional string label for trace hooks)
+    .size:          0       (this *must* be set to a valid size in bytes)
+    .type:          SG_BUFFERTYPE_VERTEXBUFFER
+    .usage:         SG_USAGE_IMMUTABLE
+    .shader_write   false
+    .content        0
+    .label          0       (optional string label for trace hooks)
 
     The dbg_label will be ignored by sokol_gfx.h, it is only useful
     when hooking into sg_make_buffer() or sg_init_buffer() via
@@ -1412,6 +1426,7 @@ typedef struct sg_buffer_desc {
     int size;
     sg_buffer_type type;
     sg_usage usage;
+    bool shader_write;
     const void* content;
     const char* label;
     /* GL specific */
@@ -2373,6 +2388,7 @@ typedef struct {
     _sg_slot_t slot;
     int size;
     int append_pos;
+    bool shader_write;
     bool append_overflow;
     sg_buffer_type type;
     sg_usage usage;
@@ -2472,6 +2488,7 @@ typedef struct {
     _sg_slot_t slot;
     int size;
     int append_pos;
+    bool shader_write;
     bool append_overflow;
     sg_buffer_type type;
     sg_usage usage;
@@ -2658,6 +2675,7 @@ typedef struct {
     _sg_slot_t slot;
     int size;
     int append_pos;
+    bool shader_write;
     bool append_overflow;
     sg_buffer_type type;
     sg_usage usage;
@@ -2665,6 +2683,8 @@ typedef struct {
     uint32_t append_frame_index;
     uint32_t map_frame_index;
     ID3D11Buffer* d3d11_buf;
+    ID3D11ShaderResourceView* d3d11_srv;
+    ID3D11UnorderedAccessView* d3d11_uav;
     int64_t used_frame;
 } _sg_buffer_t;
 
@@ -2806,7 +2826,7 @@ typedef struct {
     ID3D11Buffer* zero_cbs[SG_MAX_SHADERSTAGE_UBS];
     ID3D11ShaderResourceView* zero_srvs[SG_MAX_SHADERSTAGE_IMAGES];
     ID3D11SamplerState* zero_smps[SG_MAX_SHADERSTAGE_IMAGES];
-    ID3D11UnorderedAccessView* zero_uavs[SG_MAX_SHADERSTAGE_IMAGES_CS];
+    ID3D11UnorderedAccessView* zero_uavs[SG_MAX_SHADERSTAGE_UAVS];
     /* global subresourcedata array for texture updates */
     D3D11_SUBRESOURCE_DATA subres_data[SG_MAX_MIPMAPS * SG_MAX_TEXTUREARRAY_LAYERS];
 } _sg_d3d11_backend_t;
@@ -2864,6 +2884,7 @@ typedef struct {
     _sg_slot_t slot;
     int size;
     int append_pos;
+    bool shader_write;
     bool append_overflow;
     sg_buffer_type type;
     sg_usage usage;
@@ -3054,6 +3075,7 @@ typedef enum {
     _SG_VALIDATE_BUFFERDESC_SIZE,
     _SG_VALIDATE_BUFFERDESC_CONTENT,
     _SG_VALIDATE_BUFFERDESC_NO_CONTENT,
+    _SG_VALIDATE_BUFFERDESC_TYPE,
 
     /* image creation */
     _SG_VALIDATE_IMAGEDESC_CANARY,
@@ -3265,6 +3287,10 @@ _SOKOL_PRIVATE int _sg_uniform_size(sg_uniform_type type, int count) {
         case SG_UNIFORMTYPE_FLOAT3:     return 12 * count; /* FIXME: std140??? */
         case SG_UNIFORMTYPE_FLOAT4:     return 16 * count;
         case SG_UNIFORMTYPE_MAT4:       return 64 * count;
+        case SG_UNIFORMTYPE_INT:        return 4;
+        case SG_UNIFORMTYPE_INT2:        return 8 * count;
+        case SG_UNIFORMTYPE_INT3:        return 12 * count;
+        case SG_UNIFORMTYPE_INT4:        return 16 * count;
         default:
             SOKOL_UNREACHABLE;
             return -1;
@@ -3795,8 +3821,13 @@ _SOKOL_PRIVATE void _sg_apply_bindings(
     _sg_buffer_t** vbs, const int* vb_offsets, int num_vbs,
     _sg_buffer_t* ib, int ib_offset,
     _sg_image_t** vs_imgs, int num_vs_imgs,
-    _sg_image_t** fs_imgs, int num_fs_imgs, 
-    _sg_image_t** cs_imgs, int num_cs_imgs)
+    _sg_buffer_t** vs_bufs, int num_vs_bufs,
+    _sg_image_t** fs_imgs, int num_fs_imgs,
+    _sg_buffer_t** fs_bufs, int num_fs_bufs,
+    _sg_image_t** cs_imgs, int num_cs_imgs, 
+    _sg_buffer_t** cs_bufs, int num_cs_bufs,
+    _sg_image_t** cs_img_uavs, int num_cs_img_uavs,
+    _sg_buffer_t** cs_buf_uavs, int num_cs_buf_uavs)
 {
     SOKOL_ASSERT(pip);
     SOKOL_ASSERT(vbs && vb_offsets);
@@ -3806,8 +3837,13 @@ _SOKOL_PRIVATE void _sg_apply_bindings(
     _SOKOL_UNUSED(vbs); _SOKOL_UNUSED(vb_offsets); _SOKOL_UNUSED(num_vbs);
     _SOKOL_UNUSED(ib); _SOKOL_UNUSED(ib_offset);
     _SOKOL_UNUSED(vs_imgs); _SOKOL_UNUSED(num_vs_imgs);
+    _SOKOL_UNUSED(vs_bufs); _SOKOL_UNUSED(num_vs_bufs);
     _SOKOL_UNUSED(fs_imgs); _SOKOL_UNUSED(num_fs_imgs);
+    _SOKOL_UNUSED(fs_bufs); _SOKOL_UNUSED(num_fs_bufs);
     _SOKOL_UNUSED(cs_imgs); _SOKOL_UNUSED(num_cs_imgs);
+    _SOKOL_UNUSED(cs_bufs); _SOKOL_UNUSED(num_cs_bufs);
+    _SOKOL_UNUSED(cs_img_uavs); _SOKOL_UNUSED(num_cs_img_uavs);
+    _SOKOL_UNUSED(cs_buf_uavs); _SOKOL_UNUSED(num_cs_buf_uavs);
 }
 
 _SOKOL_PRIVATE void _sg_apply_uniforms(sg_shader_stage stage_index, int ub_index, const void* data, int num_bytes) {
@@ -6134,8 +6170,13 @@ _SOKOL_PRIVATE void _sg_apply_bindings(
     _sg_buffer_t** vbs, const int* vb_offsets, int num_vbs,
     _sg_buffer_t* ib, int ib_offset,
     _sg_image_t** vs_imgs, int num_vs_imgs,
-    _sg_image_t** fs_imgs, int num_fs_imgs, 
-	_sg_image_t** cs_imgs, int num_cs_imgs)
+    _sg_buffer_t** vs_bufs, int num_vs_bufs,
+    _sg_image_t** fs_imgs, int num_fs_imgs,
+    _sg_buffer_t** fs_bufs, int num_fs_bufs,
+	_sg_image_t** cs_imgs, int num_cs_imgs,
+    _sg_buffer_t** cs_bufs, int num_cs_bufs,
+    _sg_image_t** cs_img_uavs, int num_cs_img_uavs,
+    _sg_buffer_t** cs_buf_uavs, int num_cs_buf_uavs)
 {
     SOKOL_ASSERT(pip);
     _SOKOL_UNUSED(num_fs_imgs);
@@ -6220,11 +6261,42 @@ _SOKOL_PRIVATE void _sg_apply_bindings(
         }
     }
     _SG_GL_CHECK_ERROR();
+
+    /* TODO: vertex/fragment shader buffers */
+    if (vs_bufs && num_vs_bufs) {
+        _SOKOL_UNUSED(vs_bufs);
+        _SOKOL_UNUSED(num_vs_bufs);
+        SOKOL_ASSERT(0);
+    }
+
+    if (fs_bufs && num_fs_bufs) {
+        _SOKOL_UNUSED(fs_bufs);
+        _SOKOL_UNUSED(num_fs_bufs);
+        SOKOL_ASSERT(0);
+    }
     
-    // TODO: compute shader images */
+    /* TODO: compute shader images */
     if (cs_imgs && num_cs_imgs) {
         _SOKOL_UNUSED(cs_imgs);
         _SOKOL_UNUSED(num_cs_imgs);
+        SOKOL_ASSERT(0);
+    }
+
+    if (cs_bufs && num_cs_bufs) {
+        _SOKOL_UNUSED(cs_bufs);
+        _SOKOL_UNUSED(num_cs_bufs);
+        SOKOL_ASSERT(0);
+    }
+
+    if (cs_img_uavs && num_cs_img_uavs) {
+        _SOKOL_UNUSED(cs_img_uavs);
+        _SOKOL_UNUSED(num_cs_img_uavs);
+        SOKOL_ASSERT(0);
+    }
+
+    if (cs_buf_uavs && num_cs_buf_uavs) {
+        _SOKOL_UNUSED(cs_buf_uavs);
+        _SOKOL_UNUSED(num_cs_buf_uavs);
         SOKOL_ASSERT(0);
     }
 }
@@ -6715,6 +6787,7 @@ _SOKOL_PRIVATE void _sg_d3d11_init_caps(void) {
         info->blend  = 0 != (dxgi_fmt_caps & D3D11_FORMAT_SUPPORT_BLENDABLE);
         info->msaa   = 0 != (dxgi_fmt_caps & D3D11_FORMAT_SUPPORT_MULTISAMPLE_RENDERTARGET);
         info->depth  = 0 != (dxgi_fmt_caps & D3D11_FORMAT_SUPPORT_DEPTH_STENCIL);
+        info->uav    = 0 != (dxgi_fmt_caps & D3D11_FORMAT_SUPPORT_TYPED_UNORDERED_ACCESS_VIEW);
         if (info->depth) {
             info->render = true;
         }
@@ -6759,7 +6832,7 @@ _SOKOL_PRIVATE void _sg_d3d11_clear_state(void) {
     ID3D11DeviceContext_PSSetShaderResources(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_IMAGES, _sg.d3d11.zero_srvs);
     ID3D11DeviceContext_VSSetSamplers(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_IMAGES, _sg.d3d11.zero_smps);
     ID3D11DeviceContext_PSSetSamplers(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_IMAGES, _sg.d3d11.zero_smps);
-    ID3D11DeviceContext_CSSetUnorderedAccessViews(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_IMAGES_CS, _sg.d3d11.zero_uavs, 0);
+    ID3D11DeviceContext_CSSetUnorderedAccessViews(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_UAVS, _sg.d3d11.zero_uavs, 0);
     ID3D11DeviceContext_CSSetShaderResources(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_IMAGES, _sg.d3d11.zero_srvs);
     ID3D11DeviceContext_CSSetSamplers(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_IMAGES, _sg.d3d11.zero_smps);
 }
@@ -6790,6 +6863,7 @@ _SOKOL_PRIVATE sg_resource_state _sg_create_buffer(_sg_buffer_t* buf, const sg_b
     SOKOL_ASSERT(buf && desc);
     SOKOL_ASSERT(!buf->d3d11_buf);
     buf->size = desc->size;
+    buf->shader_write = desc->shader_write;
     buf->append_pos = 0;
     buf->append_overflow = false;
     buf->type = desc->type;
@@ -6803,11 +6877,25 @@ _SOKOL_PRIVATE sg_resource_state _sg_create_buffer(_sg_buffer_t* buf, const sg_b
         ID3D11Buffer_AddRef(buf->d3d11_buf);
     }
     else {
+        UINT bind_flags;
+        switch (buf->type) {
+            case SG_BUFFERTYPE_VERTEXBUFFER:    bind_flags = D3D11_BIND_VERTEX_BUFFER; break;
+            case SG_BUFFERTYPE_INDEXBUFFER:     bind_flags = D3D11_BIND_INDEX_BUFFER; break;
+            case SG_BUFFERTYPE_RAW:             bind_flags = D3D11_BIND_SHADER_RESOURCE; break;
+            default:
+                SOKOL_UNREACHABLE;
+                break;
+        }
+
+        if (desc->shader_write) {
+            bind_flags |= D3D11_BIND_UNORDERED_ACCESS;
+        }
+
         D3D11_BUFFER_DESC d3d11_desc;
         memset(&d3d11_desc, 0, sizeof(d3d11_desc));
         d3d11_desc.ByteWidth = buf->size;
         d3d11_desc.Usage = _sg_d3d11_usage(buf->usage);
-        d3d11_desc.BindFlags = buf->type == SG_BUFFERTYPE_VERTEXBUFFER ? D3D11_BIND_VERTEX_BUFFER : D3D11_BIND_INDEX_BUFFER;
+        d3d11_desc.BindFlags = bind_flags;
         d3d11_desc.CPUAccessFlags = _sg_d3d11_cpu_access_flags(buf->usage);
         D3D11_SUBRESOURCE_DATA* init_data_ptr = 0;
         D3D11_SUBRESOURCE_DATA init_data;
@@ -6817,9 +6905,39 @@ _SOKOL_PRIVATE sg_resource_state _sg_create_buffer(_sg_buffer_t* buf, const sg_b
             init_data.pSysMem = desc->content;
             init_data_ptr = &init_data;
         }
+        else if (desc->content && buf->type == SG_BUFFERTYPE_RAW) {
+            init_data.pSysMem = desc->content;
+            init_data_ptr = &init_data;            
+        }
+
+        if (buf->type == SG_BUFFERTYPE_RAW) {
+            d3d11_desc.MiscFlags = D3D11_RESOURCE_MISC_BUFFER_ALLOW_RAW_VIEWS;
+        }
+
         HRESULT hr = ID3D11Device_CreateBuffer(_sg.d3d11.dev, &d3d11_desc, init_data_ptr, &buf->d3d11_buf);
         _SOKOL_UNUSED(hr);
         SOKOL_ASSERT(SUCCEEDED(hr) && buf->d3d11_buf);
+
+        /* create views SRV/UAV */
+        if (buf->type == SG_BUFFERTYPE_RAW) {
+            D3D11_SHADER_RESOURCE_VIEW_DESC srv_desc = { 0 };
+            srv_desc.ViewDimension = D3D11_SRV_DIMENSION_BUFFEREX;
+            srv_desc.Format = DXGI_FORMAT_R32_TYPELESS;
+            srv_desc.BufferEx.Flags = D3D11_BUFFEREX_SRV_FLAG_RAW;
+            srv_desc.BufferEx.NumElements = buf->size / 4;
+            hr = ID3D11Device_CreateShaderResourceView(_sg.d3d11.dev, (ID3D11Resource*)buf->d3d11_buf, &srv_desc, &buf->d3d11_srv);
+            SOKOL_ASSERT(SUCCEEDED(hr) && buf->d3d11_srv);
+
+            if (desc->shader_write) {
+                D3D11_UNORDERED_ACCESS_VIEW_DESC uav_desc = { 0 };
+                uav_desc.ViewDimension = D3D11_UAV_DIMENSION_BUFFER;
+                uav_desc.Format = DXGI_FORMAT_R32_TYPELESS;
+                uav_desc.Buffer.Flags = D3D11_BUFFER_UAV_FLAG_RAW;
+                uav_desc.Buffer.NumElements = buf->size / 4;
+                hr = ID3D11Device_CreateUnorderedAccessView(_sg.d3d11.dev, (ID3D11Resource*)buf->d3d11_buf, &uav_desc, &buf->d3d11_uav);
+                SOKOL_ASSERT(SUCCEEDED(hr) && buf->d3d11_uav);
+            }
+        }
     }
     return SG_RESOURCESTATE_VALID;
 }
@@ -6828,6 +6946,12 @@ _SOKOL_PRIVATE void _sg_destroy_buffer(_sg_buffer_t* buf) {
     SOKOL_ASSERT(buf);
     if (buf->d3d11_buf) {
         ID3D11Buffer_Release(buf->d3d11_buf);
+    }
+    if (buf->d3d11_srv) {
+        ID3D11Buffer_Release(buf->d3d11_srv);
+    }
+    if (buf->d3d11_uav) {
+        ID3D11Buffer_Release(buf->d3d11_uav);
     }
 }
 
@@ -7197,11 +7321,13 @@ _SOKOL_PRIVATE sg_resource_state _sg_create_shader(_sg_shader_t* shd, const sg_s
     /* shader stage uniform blocks and image slots */
     for (int stage_index = 0; stage_index < SG_NUM_SHADER_STAGES; stage_index++) {
         const sg_shader_stage_desc* stage_desc;
-        if (stage_index == SG_SHADERSTAGE_CS) {
-            stage_desc = &desc->cs;
-        } 
-        else {
-            stage_desc = (stage_index == SG_SHADERSTAGE_VS) ? &desc->vs : &desc->fs;
+        switch (stage_index) {
+            case SG_SHADERSTAGE_VS:     stage_desc = &desc->vs; break;
+            case SG_SHADERSTAGE_FS:     stage_desc = &desc->fs; break;
+            case SG_SHADERSTAGE_CS:     stage_desc = &desc->cs; break;
+            default:                    
+                SOKOL_UNREACHABLE;
+                break;
         }
         _sg_shader_stage_t* stage = &shd->stage[stage_index];
         SOKOL_ASSERT(stage->num_uniform_blocks == 0);
@@ -7749,8 +7875,13 @@ _SOKOL_PRIVATE void _sg_apply_bindings(
     _sg_buffer_t** vbs, const int* vb_offsets, int num_vbs,
     _sg_buffer_t* ib, int ib_offset,
     _sg_image_t** vs_imgs, int num_vs_imgs,
+    _sg_buffer_t** vs_bufs, int num_vs_bufs,
     _sg_image_t** fs_imgs, int num_fs_imgs,
-    _sg_image_t** cs_imgs, int num_cs_imgs)
+    _sg_buffer_t** fs_bufs, int num_fs_bufs,
+    _sg_image_t** cs_imgs, int num_cs_imgs,
+    _sg_buffer_t** cs_bufs, int num_cs_bufs,
+    _sg_image_t** cs_img_uavs, int num_cs_img_uavs,
+    _sg_buffer_t** cs_buf_uavs, int num_cs_buf_uavs)
 {
     SOKOL_ASSERT(pip);
     SOKOL_ASSERT(_sg.d3d11.ctx);
@@ -7758,73 +7889,76 @@ _SOKOL_PRIVATE void _sg_apply_bindings(
 
     /* gather all the D3D11 resources into arrays */
     ID3D11Buffer* d3d11_ib = ib ? ib->d3d11_buf : 0;
-    ID3D11Buffer* d3d11_vbs[SG_MAX_SHADERSTAGE_BUFFERS];
-    UINT d3d11_vb_offsets[SG_MAX_SHADERSTAGE_BUFFERS];
-    ID3D11ShaderResourceView* d3d11_vs_srvs[SG_MAX_SHADERSTAGE_IMAGES];
-    ID3D11SamplerState* d3d11_vs_smps[SG_MAX_SHADERSTAGE_IMAGES];
-    ID3D11ShaderResourceView* d3d11_fs_srvs[SG_MAX_SHADERSTAGE_IMAGES];
-    ID3D11SamplerState* d3d11_fs_smps[SG_MAX_SHADERSTAGE_IMAGES];
-
-    ID3D11ShaderResourceView* d3d11_cs_srvs[SG_MAX_SHADERSTAGE_IMAGES];
-    ID3D11UnorderedAccessView* d3d11_cs_uavs[SG_MAX_SHADERSTAGE_IMAGES_CS];
-    ID3D11SamplerState* d3d11_cs_smps[SG_MAX_SHADERSTAGE_IMAGES];
+    ID3D11Buffer* d3d11_vbs[SG_MAX_SHADERSTAGE_BUFFERS] = { 0 };
+    UINT d3d11_vb_offsets[SG_MAX_SHADERSTAGE_BUFFERS] = { 0 };
+    ID3D11ShaderResourceView* d3d11_vs_srvs[SG_MAX_SHADERSTAGE_IMAGES] = { 0 };
+    ID3D11SamplerState* d3d11_vs_smps[SG_MAX_SHADERSTAGE_IMAGES] = { 0 };
+    ID3D11ShaderResourceView* d3d11_fs_srvs[SG_MAX_SHADERSTAGE_IMAGES] = { 0 };
+    ID3D11SamplerState* d3d11_fs_smps[SG_MAX_SHADERSTAGE_IMAGES] = { 0 };
+    ID3D11ShaderResourceView* d3d11_cs_srvs[SG_MAX_SHADERSTAGE_IMAGES] = { 0 };
+    ID3D11UnorderedAccessView* d3d11_cs_uavs[SG_MAX_SHADERSTAGE_UAVS] = { 0 };
+    ID3D11SamplerState* d3d11_cs_smps[SG_MAX_SHADERSTAGE_IMAGES] = { 0 };
 
     int i;
+    int vs_srv_index, fs_srv_index, cs_srv_index, cs_uav_index;
+
     for (i = 0; i < num_vbs; i++) {
         SOKOL_ASSERT(vbs[i]->d3d11_buf);
         d3d11_vbs[i] = vbs[i]->d3d11_buf;
         d3d11_vb_offsets[i] = vb_offsets[i];
     }
-    for (; i < SG_MAX_SHADERSTAGE_BUFFERS; i++) {
-        d3d11_vbs[i] = 0;
-        d3d11_vb_offsets[i] = 0;
-    }
-    for (i = 0; i < num_vs_imgs; i++) {
+
+    for (i = 0, vs_srv_index = 0; i < num_vs_imgs; i++, vs_srv_index++) {
         SOKOL_ASSERT(vs_imgs[i]->d3d11_srv);
         SOKOL_ASSERT(vs_imgs[i]->d3d11_smp);
-        d3d11_vs_srvs[i] = vs_imgs[i]->d3d11_srv;
-        d3d11_vs_smps[i] = vs_imgs[i]->d3d11_smp;
+        d3d11_vs_srvs[vs_srv_index] = vs_imgs[i]->d3d11_srv;
+        d3d11_vs_smps[vs_srv_index] = vs_imgs[i]->d3d11_smp;
     }
-    for (; i < SG_MAX_SHADERSTAGE_IMAGES; i++) {
-        d3d11_vs_srvs[i] = 0;
-        d3d11_vs_smps[i] = 0;
+
+    for (i = 0; i < num_vs_bufs; i++, vs_srv_index++) {
+        SOKOL_ASSERT(vs_bufs[i]->d3d11_srv);
+        d3d11_vs_srvs[vs_srv_index] = vs_bufs[i]->d3d11_srv;
     }
-    for (i = 0; i < num_fs_imgs; i++) {
+
+    for (i = 0, fs_srv_index = 0; i < num_fs_imgs; i++, fs_srv_index++) {
         SOKOL_ASSERT(fs_imgs[i]->d3d11_srv);
         SOKOL_ASSERT(fs_imgs[i]->d3d11_smp);
-        d3d11_fs_srvs[i] = fs_imgs[i]->d3d11_srv;
-        d3d11_fs_smps[i] = fs_imgs[i]->d3d11_smp;
-    }
-    for (; i < SG_MAX_SHADERSTAGE_IMAGES; i++) {
-        d3d11_fs_srvs[i] = 0;
-        d3d11_fs_smps[i] = 0;
+        d3d11_fs_srvs[fs_srv_index] = fs_imgs[i]->d3d11_srv;
+        d3d11_fs_smps[fs_srv_index] = fs_imgs[i]->d3d11_smp;
     }
 
-	for (i = 0; i < num_cs_imgs; i++) {
+    for (i = 0; i < num_fs_bufs; i++, fs_srv_index++) {
+        SOKOL_ASSERT(fs_bufs[i]->d3d11_srv);
+        d3d11_fs_srvs[fs_srv_index] = fs_bufs[i]->d3d11_srv;
+    }
+
+	for (i = 0, cs_srv_index = 0; i < num_cs_imgs; i++, cs_srv_index++) {
         SOKOL_ASSERT(cs_imgs[i]->d3d11_srv);
         SOKOL_ASSERT(cs_imgs[i]->d3d11_smp);
-        d3d11_cs_uavs[i] = cs_imgs[i]->d3d11_uav;
-        if (!d3d11_cs_uavs[i]) {
-            d3d11_cs_srvs[i] = cs_imgs[i]->d3d11_srv;
-            d3d11_cs_smps[i] = cs_imgs[i]->d3d11_smp;
-        } 
-        else {
-            d3d11_cs_srvs[i] = 0;
-            d3d11_cs_smps[i] = 0;
-        }
+        SOKOL_ASSERT(cs_srv_index < SG_MAX_SHADERSTAGE_IMAGES);
+        d3d11_cs_srvs[cs_srv_index] = cs_imgs[i]->d3d11_srv;
+        d3d11_cs_smps[cs_srv_index] = cs_imgs[i]->d3d11_smp;
     }
 
-    int i_bak = i;
-
-    for (; i < SG_MAX_SHADERSTAGE_IMAGES_CS; i++) {
-        d3d11_cs_uavs[i] = 0;
+    for (i = 0; i < num_cs_bufs; i++, cs_srv_index++) {
+        SOKOL_ASSERT(cs_bufs[i]->d3d11_srv);
+        SOKOL_ASSERT(cs_srv_index < SG_MAX_SHADERSTAGE_IMAGES);
+        d3d11_cs_srvs[cs_srv_index] = cs_bufs[i]->d3d11_srv;
     }
 
-    for (i = i_bak; i < SG_MAX_SHADERSTAGE_IMAGES; i++) {
-        d3d11_cs_srvs[i] = 0;
-        d3d11_cs_smps[i] = 0;
+    /* uavs */
+    for (i = 0, cs_uav_index = 0; i < num_cs_img_uavs; i++, cs_uav_index++) {
+        SOKOL_ASSERT(cs_img_uavs[i]->d3d11_uav);
+        SOKOL_ASSERT(cs_uav_index < SG_MAX_SHADERSTAGE_UAVS);
+        d3d11_cs_uavs[cs_uav_index] = cs_img_uavs[i]->d3d11_uav;
     }
-    
+
+    for (i = 0; i < num_cs_buf_uavs; i++, cs_uav_index++) {
+        SOKOL_ASSERT(cs_buf_uavs[i]->d3d11_uav);
+        SOKOL_ASSERT(cs_uav_index < SG_MAX_SHADERSTAGE_UAVS);
+        d3d11_cs_uavs[cs_uav_index] = cs_buf_uavs[i]->d3d11_uav;
+    }
+
     if (!pip->shader->compute_shader) {
         ID3D11DeviceContext_IASetVertexBuffers(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_BUFFERS, d3d11_vbs, pip->d3d11_vb_strides, d3d11_vb_offsets);
         ID3D11DeviceContext_IASetIndexBuffer(_sg.d3d11.ctx, d3d11_ib, pip->d3d11_index_format, ib_offset);
@@ -7834,7 +7968,7 @@ _SOKOL_PRIVATE void _sg_apply_bindings(
         ID3D11DeviceContext_PSSetSamplers(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_IMAGES, d3d11_fs_smps);
     }
     else {
-	    ID3D11DeviceContext_CSSetUnorderedAccessViews(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_IMAGES_CS, d3d11_cs_uavs, 0);
+	    ID3D11DeviceContext_CSSetUnorderedAccessViews(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_UAVS, d3d11_cs_uavs, 0);
         ID3D11DeviceContext_CSSetShaderResources(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_IMAGES, d3d11_cs_srvs);
         ID3D11DeviceContext_CSSetSamplers(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_IMAGES, d3d11_cs_smps);
     }
@@ -7843,7 +7977,7 @@ _SOKOL_PRIVATE void _sg_apply_bindings(
 
 _SOKOL_PRIVATE void _sg_apply_uniforms(sg_shader_stage stage_index, int ub_index, const void* data, int num_bytes) {
     _SOKOL_UNUSED(num_bytes);
-    SOKOL_ASSERT(_sg.d3d11.ctx && _sg.d3d11.in_pass);
+    SOKOL_ASSERT(_sg.d3d11.ctx && (_sg.d3d11.in_pass || stage_index == SG_SHADERSTAGE_CS));
     SOKOL_ASSERT(data && (num_bytes > 0));
     SOKOL_ASSERT((stage_index >= 0) && ((int)stage_index < SG_NUM_SHADER_STAGES));
     SOKOL_ASSERT((ub_index >= 0) && (ub_index < SG_MAX_SHADERSTAGE_UBS));
@@ -7882,7 +8016,7 @@ _SOKOL_PRIVATE void _sg_dispatch(int thread_group_x, int thread_group_y, int thr
 
     /* cleanup */
     /* TODO: put this into a better place */
-    ID3D11DeviceContext_CSSetUnorderedAccessViews(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_IMAGES_CS, _sg.d3d11.zero_uavs, 0);
+    ID3D11DeviceContext_CSSetUnorderedAccessViews(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_UAVS, _sg.d3d11.zero_uavs, 0);
     ID3D11DeviceContext_CSSetShaderResources(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_IMAGES, _sg.d3d11.zero_srvs);
     ID3D11DeviceContext_CSSetSamplers(_sg.d3d11.ctx, 0, SG_MAX_SHADERSTAGE_IMAGES, _sg.d3d11.zero_smps);
 }
@@ -9517,8 +9651,13 @@ _SOKOL_PRIVATE void _sg_apply_bindings(
     _sg_buffer_t** vbs, const int* vb_offsets, int num_vbs,
     _sg_buffer_t* ib, int ib_offset,
     _sg_image_t** vs_imgs, int num_vs_imgs,
+    _sg_buffer_t** vs_bufs, int num_vs_bufs,
     _sg_image_t** fs_imgs, int num_fs_imgs,
-    _sg_image_t** cs_imgs, int num_cs_imgs)
+    _sg_buffer_t** fs_bufs, int num_fs_bufs,
+    _sg_image_t** cs_imgs, int num_cs_imgs,
+    _sg_buffer_t** cs_bufs, int num_cs_bufs,
+    _sg_image_t** cs_img_uavs, int num_cs_img_uavs,
+    _sg_buffer_t** cs_buf_uavs, int num_cs_buf_uavs)
 {
     SOKOL_ASSERT(_sg.mtl.in_pass);
     if (!_sg.mtl.pass_valid) {
@@ -9583,10 +9722,41 @@ _SOKOL_PRIVATE void _sg_apply_bindings(
         }
     }
 
+    /* TODO: apply vertex/fragment shader buffers */
+    if (vs_bufs && num_vs_bufs) {
+        _SOKOL_UNUSED(vs_bufs);
+        _SOKOL_UNUSED(num_vs_bufs);
+        SOKOL_ASSERT(0);
+    }
+
+    if (fs_bufs && num_fs_bufs) {
+        _SOKOL_UNUSED(fs_bufs);
+        _SOKOL_UNUSED(num_fs_bufs);
+        SOKOL_ASSERT(0);
+    }
+
     /* TODO: apply compute shader images */
     if (cs_imgs && num_cs_imgs > 0) {
         _SOKOL_UNUSED(cs_imgs);
         _SOKOL_UNUSED(num_cs_imgs);
+        SOKOL_ASSERT(0);
+    }
+
+    if (cs_bufs && num_cs_bufs > 0) {
+        _SOKOL_UNUSED(cs_bufs);
+        _SOKOL_UNUSED(num_cs_bufs);
+        SOKOL_ASSERT(0);
+    }
+
+    if (cs_img_uavs && num_cs_img_uavs > 0) {
+        _SOKOL_UNUSED(cs_img_uavs);
+        _SOKOL_UNUSED(num_cs_img_uavs);
+        SOKOL_ASSERT(0);
+    }
+
+    if (cs_buf_uavs && num_cs_buf_uavs > 0) {
+        _SOKOL_UNUSED(cs_buf_uavs);
+        _SOKOL_UNUSED(num_cs_buf_uavs);
         SOKOL_ASSERT(0);
     }
 }
@@ -10052,6 +10222,7 @@ _SOKOL_PRIVATE const char* _sg_validate_string(_sg_validate_error_t err) {
         case _SG_VALIDATE_BUFFERDESC_SIZE:          return "sg_buffer_desc.size cannot be 0";
         case _SG_VALIDATE_BUFFERDESC_CONTENT:       return "immutable buffers must be initialized with content (sg_buffer_desc.content)";
         case _SG_VALIDATE_BUFFERDESC_NO_CONTENT:    return "dynamic/stream usage buffers cannot be initialized with content";
+        case _SG_VALIDATE_BUFFERDESC_TYPE:          return "buffer storage cannot be vertex or index type";
 
         /* image creation validation errros */
         case _SG_VALIDATE_IMAGEDESC_CANARY:             return "sg_image_desc not initialized";
@@ -10212,12 +10383,18 @@ _SOKOL_PRIVATE bool _sg_validate_buffer_desc(const sg_buffer_desc* desc) {
         SOKOL_VALIDATE(desc->_end_canary == 0, _SG_VALIDATE_BUFFERDESC_CANARY);
         SOKOL_VALIDATE(desc->size > 0, _SG_VALIDATE_BUFFERDESC_SIZE);
         bool ext = (0 != desc->gl_buffers[0]) || (0 != desc->mtl_buffers[0]) || (0 != desc->d3d11_buffer);
-        if (!ext && (desc->usage == SG_USAGE_IMMUTABLE)) {
-            SOKOL_VALIDATE(0 != desc->content, _SG_VALIDATE_BUFFERDESC_CONTENT);
+        if (desc->usage != _SG_USAGE_DEFAULT) {
+            if (!ext && (desc->usage == SG_USAGE_IMMUTABLE)) {
+                SOKOL_VALIDATE(0 != desc->content, _SG_VALIDATE_BUFFERDESC_CONTENT);
+            }
+            else {
+                SOKOL_VALIDATE(0 == desc->content, _SG_VALIDATE_BUFFERDESC_NO_CONTENT);
+            }
         }
-        else {
-            SOKOL_VALIDATE(0 == desc->content, _SG_VALIDATE_BUFFERDESC_NO_CONTENT);
+        if (desc->shader_write) {
+            SOKOL_VALIDATE(desc->type != SG_BUFFERTYPE_VERTEXBUFFER && desc->type != SG_BUFFERTYPE_INDEXBUFFER, _SG_VALIDATE_BUFFERDESC_TYPE);
         }
+
         return SOKOL_VALIDATE_END();
     #endif
 }
@@ -10594,7 +10771,7 @@ _SOKOL_PRIVATE bool _sg_validate_apply_bindings(const sg_bindings* bindings) {
                 const _sg_buffer_t* buf = _sg_lookup_buffer(&_sg.pools, bindings->vertex_buffers[i].id);
                 SOKOL_VALIDATE(buf != 0, _SG_VALIDATE_ABND_VB_EXISTS);
                 if (buf && buf->slot.state == SG_RESOURCESTATE_VALID) {
-                    SOKOL_VALIDATE(SG_BUFFERTYPE_VERTEXBUFFER == buf->type, _SG_VALIDATE_ABND_VB_TYPE);
+                    SOKOL_VALIDATE(SG_BUFFERTYPE_VERTEXBUFFER == buf->type || SG_BUFFERTYPE_RAW == buf->type, _SG_VALIDATE_ABND_VB_TYPE);
                     SOKOL_VALIDATE(!buf->append_overflow, _SG_VALIDATE_ABND_VB_OVERFLOW);
                 }
             }
@@ -10682,7 +10859,7 @@ _SOKOL_PRIVATE bool _sg_validate_apply_uniforms(sg_shader_stage stage_index, int
         _SOKOL_UNUSED(num_bytes);
         return true;
     #else
-        SOKOL_ASSERT((stage_index == SG_SHADERSTAGE_VS) || (stage_index == SG_SHADERSTAGE_FS));
+        SOKOL_ASSERT((stage_index == SG_SHADERSTAGE_VS) || (stage_index == SG_SHADERSTAGE_FS) || (stage_index == SG_SHADERSTAGE_CS));
         SOKOL_ASSERT((ub_index >= 0) && (ub_index < SG_MAX_SHADERSTAGE_UBS));
         SOKOL_VALIDATE_BEGIN();
         SOKOL_VALIDATE(_sg.cur_pipeline.id != SG_INVALID_ID, _SG_VALIDATE_AUB_NO_PIPELINE);
@@ -10765,8 +10942,8 @@ _SOKOL_PRIVATE bool _sg_validate_update_image(const _sg_image_t* img, const sg_i
 /*== fill in desc default values =============================================*/
 _SOKOL_PRIVATE sg_buffer_desc _sg_buffer_desc_defaults(const sg_buffer_desc* desc) {
     sg_buffer_desc def = *desc;
-    def.type = _sg_def(def.type, SG_BUFFERTYPE_VERTEXBUFFER);
-    def.usage = _sg_def(def.usage, SG_USAGE_IMMUTABLE);
+    def.type =  _sg_def(def.type, SG_BUFFERTYPE_VERTEXBUFFER);
+    def.usage = !def.shader_write ? _sg_def(def.usage, SG_USAGE_IMMUTABLE) : _sg_def(def.usage, _SG_USAGE_DEFAULT);
     return def;
 }
 
@@ -11638,6 +11815,19 @@ SOKOL_API_IMPL void sg_apply_bindings(const sg_bindings* bindings) {
         }
     }
 
+    _sg_buffer_t* vs_bufs[SG_MAX_SHADERSTAGE_BUFFERS] = { 0 };
+    int num_vs_bufs = 0;
+    for (int i = 0; i < SG_MAX_SHADERSTAGE_BUFFERS; i++, num_vs_bufs++) {
+        if (bindings->vs_buffers[i].id) {
+            vs_bufs[i] = _sg_lookup_buffer(&_sg.pools, bindings->vs_buffers[i].id);
+            SOKOL_ASSERT(vs_bufs[i]);
+            _sg.next_draw_valid &= (SG_RESOURCESTATE_VALID == vs_bufs[i]->slot.state);
+        }
+        else {
+            break;
+        }
+    }
+
     _sg_image_t* fs_imgs[SG_MAX_SHADERSTAGE_IMAGES] = { 0 };
     int num_fs_imgs = 0;
     for (int i = 0; i < SG_MAX_SHADERSTAGE_IMAGES; i++, num_fs_imgs++) {
@@ -11646,6 +11836,19 @@ SOKOL_API_IMPL void sg_apply_bindings(const sg_bindings* bindings) {
             SOKOL_ASSERT(fs_imgs[i]);
             _sg.next_draw_valid &= (SG_RESOURCESTATE_VALID == fs_imgs[i]->slot.state);
         } 
+        else {
+            break;
+        }
+    }
+
+    _sg_buffer_t* fs_bufs[SG_MAX_SHADERSTAGE_BUFFERS] = { 0 };
+    int num_fs_bufs = 0;
+    for (int i = 0; i < SG_MAX_SHADERSTAGE_BUFFERS; i++, num_fs_bufs++) {
+        if (bindings->fs_buffers[i].id) {
+            fs_bufs[i] = _sg_lookup_buffer(&_sg.pools, bindings->fs_buffers[i].id);
+            SOKOL_ASSERT(vs_bufs[i]);
+            _sg.next_draw_valid &= (SG_RESOURCESTATE_VALID == fs_bufs[i]->slot.state);
+        }
         else {
             break;
         }
@@ -11664,10 +11867,60 @@ SOKOL_API_IMPL void sg_apply_bindings(const sg_bindings* bindings) {
         }
     }
 
+    _sg_buffer_t* cs_bufs[SG_MAX_SHADERSTAGE_BUFFERS] = { 0 };
+    int num_cs_bufs = 0;
+
+    for (int i = 0; i < SG_MAX_SHADERSTAGE_BUFFERS; i++, num_cs_bufs++) {
+        if (bindings->cs_buffers[i].id) {
+            cs_bufs[i] = _sg_lookup_buffer(&_sg.pools, bindings->cs_buffers[i].id);
+            SOKOL_ASSERT(cs_bufs[i]);
+            _sg.next_draw_valid &= (SG_RESOURCESTATE_VALID == cs_bufs[i]->slot.state);
+        }
+        else {
+            break;
+        }
+    }
+
+    _sg_image_t* cs_img_uavs[SG_MAX_SHADERSTAGE_UAVS] = { 0 };
+    int num_cs_img_uavs = 0;
+    for (int i = 0; i < SG_MAX_SHADERSTAGE_UAVS; i++, num_cs_img_uavs++) {
+        if (bindings->cs_image_uavs[i].id) {
+            cs_img_uavs[i] = _sg_lookup_image(&_sg.pools, bindings->cs_image_uavs[i].id);
+            SOKOL_ASSERT(cs_img_uavs[i]);
+            SOKOL_ASSERT(cs_img_uavs[i]->shader_write);
+            _sg.next_draw_valid &= (SG_RESOURCESTATE_VALID == cs_img_uavs[i]->slot.state);
+        }
+        else {
+            break;
+        }
+    }
+
+    _sg_buffer_t* cs_buf_uavs[SG_MAX_SHADERSTAGE_UAVS] = { 0 };
+    int num_cs_buf_uavs = 0;
+    for (int i = 0; i < SG_MAX_SHADERSTAGE_UAVS; i++, num_cs_buf_uavs++) {
+        if (bindings->cs_buffer_uavs[i].id) {
+            cs_buf_uavs[i] = _sg_lookup_buffer(&_sg.pools, bindings->cs_buffer_uavs[i].id);
+            SOKOL_ASSERT(cs_buf_uavs[i]);
+            SOKOL_ASSERT(cs_buf_uavs[i]->shader_write);
+            _sg.next_draw_valid &= (SG_RESOURCESTATE_VALID == cs_buf_uavs[i]->slot.state);
+        }
+        else {
+            break;
+        }
+    }
+
     if (_sg.next_draw_valid) {
         const int* vb_offsets = bindings->vertex_buffer_offsets;
         int ib_offset = bindings->index_buffer_offset;
-        _sg_apply_bindings(pip, vbs, vb_offsets, num_vbs, ib, ib_offset, vs_imgs, num_vs_imgs, fs_imgs, num_fs_imgs, cs_imgs, num_cs_imgs);
+        _sg_apply_bindings(pip, vbs, vb_offsets, num_vbs, ib, ib_offset, 
+            vs_imgs, num_vs_imgs, 
+            vs_bufs, num_vs_bufs,
+            fs_imgs, num_fs_imgs,
+            fs_bufs, num_fs_bufs,
+            cs_imgs, num_cs_imgs, 
+            cs_bufs, num_cs_bufs, 
+            cs_img_uavs, num_cs_img_uavs,
+            cs_buf_uavs, num_cs_buf_uavs);
         _SG_TRACE_ARGS(apply_bindings, bindings);
     }
     else {
@@ -11677,7 +11930,7 @@ SOKOL_API_IMPL void sg_apply_bindings(const sg_bindings* bindings) {
 
 SOKOL_API_IMPL void sg_apply_uniforms(sg_shader_stage stage, int ub_index, const void* data, int num_bytes) {
     SOKOL_ASSERT(_sg.valid);
-    SOKOL_ASSERT((stage == SG_SHADERSTAGE_VS) || (stage == SG_SHADERSTAGE_FS));
+    SOKOL_ASSERT((stage == SG_SHADERSTAGE_VS) || (stage == SG_SHADERSTAGE_FS) || (stage == SG_SHADERSTAGE_CS));
     SOKOL_ASSERT((ub_index >= 0) && (ub_index < SG_MAX_SHADERSTAGE_UBS));
     SOKOL_ASSERT(data && (num_bytes > 0));
     if (!_sg_validate_apply_uniforms(stage, ub_index, data, num_bytes)) {
@@ -11685,7 +11938,7 @@ SOKOL_API_IMPL void sg_apply_uniforms(sg_shader_stage stage, int ub_index, const
         _SG_TRACE_NOARGS(err_draw_invalid);
         return;
     }
-    if (!_sg.pass_valid) {
+    if (!_sg.pass_valid && stage != SG_SHADERSTAGE_CS) {
         _SG_TRACE_NOARGS(err_pass_invalid);
         return;
     }