+++ CHANGES.md
@@ -24,6 +24,7 @@
 - [x] Raspberry build
 - [ ] iOS build
 - [x] EXPERIMENTAL: D3D11 compute-shader support
+- [ ] EXPERIMENTAL: sokol debugging support for compute-shader
 
 ### Changed
 - [x] BREAKING - gfx: immediate and staged APIs are now the same. creation/destroy/query functions 
@@ -33,6 +34,7 @@
 - [x] Deferred destroys. Now objects in use will never be destroyed even if the user calls _destroy_
 - [ ] Error handling for missing shaders
 - [x] immediate mode api conflicts with imgui
+- [ ] fixed sx_matx_inv 
 
 ## [0.1]:
 ### Added